requirements.txt
fastapi~=0.111.0
uvicorn~=0.30.1
redis~=5.0.4
python-dotenv~=1.0.1
pydantic>=2.7.3
starlette~=0.37.2
pydantic-settings>=2.3.1
ruff
httpx~=0.27.0
pycountry
pydantic-extra-types
pytest>=8.2.2
pytest-asyncio>=0.23.7
async-asgi-testclient>=1.4.11
